Force reqiure authentication
